+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Reset list title to "Wall of Names" if title is blank
 - Remove " - " prefix from document title if list title is blank
+- Randomise size of spawned sharks
 
 ## [1.1.4] - 2025-05-12